change int to float

+++ Python/PythonCore/HomeWork/HomeWork_2_20200609/Task3.py
@@ -10,6 +10,6 @@ Created on Tue Jun  9 20:14:51 2020
 width = input("Enter width:\n")
 height = input("Enter height:\n")
 
-print("Area of rectangle is: ", int(width) * int(height))
-print("Perimeter of rectangle is: ", 2*(int(width) + int(height)))
+print("Area of rectangle is: ", float(width) * float(height))
+print("Perimeter of rectangle is: ", 2*(float(width) + float(height)))
